requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.2
djangorestframework==3.13.1
idna==3.3
kombu==5.2.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
prompt-toolkit==3.0.27
pyparsing==3.0.7
pytz==2021.3
redis==3.5.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
